Removed RepetitiveBySection dependencies: should be replaced by RepetitiveBySectionReMaster.kt on later release
