+++ src/UnifyMain/kotlin/unify/parsers/ArgumentsParser.kt
@@ -2,7 +2,6 @@ package unify.parsers
 
 import language.parsers.ParserStatic
 import language.sections.OptionalSection
-import language.sections.RepetitiveBySection
 import language.sections.RepetitiveBySectionReMaster
 import unify.tokens.characters.Coma
 import unify.tokens.characters.LBracket
