+++ src/UnifyMain/kotlin/unify/parsers/expressions/TFunctionCallParser.kt
@@ -1,0 +1,16 @@
+package unify.parsers.expressions
+
+import language.parsers.ParserStatic
+import language.sections.AlternativeSection
+import lib.matcher.TestableStatic
+import unify.parsers.TerminatorParser
+import unify.parsers.functions.CallParser
+
+
+class TFunctionCallParser : ParserStatic(
+    AlternativeSection(
+        ParameterFunctionCall(),
+        InfixFunctionCallParser()
+    ),
+    TerminatorParser()
+)
