+++ src/UnifyMain/kotlin/unify/parsers/functions/SimpleParameterParser.kt
@@ -1,16 +1,13 @@
 package unify.parsers.functions
 
 import language.parsers.ParserStatic
-import language.sections.OptionalSection
-import language.sections.RepetitiveBySection
-import language.sections.ZeroOrMany
+import language.sections.RepetitiveBySectionReMaster
 import unify.parsers.expressions.ExpressionParser
 import unify.tokens.characters.Coma
 
 class SimpleParameterParser : ParserStatic(
-        RepetitiveBySection(
-                ExpressionParser(),
-                Coma
-
-        )
+    RepetitiveBySectionReMaster(
+        Coma,
+        ExpressionParser()
+    )
 )
