+++ src/UnifyMain/kotlin/unify/parsers/literals/ArrayParser.kt
@@ -2,21 +2,21 @@ package unify.parsers.literals
 
 import language.parsers.ParserStatic
 import language.sections.OptionalSection
-import language.sections.RepetitiveBySection
+import language.sections.RepetitiveBySectionReMaster
 import unify.parsers.expressions.ExpressionParser
 import unify.tokens.characters.Coma
 import unify.tokens.characters.LSBracket
 import unify.tokens.characters.RSBracket
 
 class ArrayParser : ParserStatic(
-        LSBracket,
-        OptionalSection(
-                RepetitiveBySection(
-                        ExpressionParser(),
-                        Coma
-                )
-        ),
-        RSBracket
+    LSBracket,
+    OptionalSection(
+        RepetitiveBySectionReMaster(
+            Coma,
+            ExpressionParser()
+        )
+    ),
+    RSBracket
 ) {
-    override val TAG ="ArrayParser"
+    override val TAG = "ArrayParser"
 }
