+++ src/UnifyMain/kotlin/unify/parsers/objects/EnumBodyParser.kt
@@ -2,7 +2,7 @@ package unify.parsers.objects
 
 import language.parsers.ParserStatic
 import language.sections.OptionalSection
-import language.sections.RepetitiveBySection
+import language.sections.RepetitiveBySectionReMaster
 import unify.tokens.characters.Coma
 import unify.tokens.characters.LBrace
 import unify.tokens.characters.RBrace
@@ -10,12 +10,12 @@ import unify.tokens.strings.Identifier
 
 
 class EnumBodyParser : ParserStatic(
-        LBrace,
-        OptionalSection(
-                RepetitiveBySection(
-                        Identifier,
-                        Coma
-                )
-        ),
-        RBrace
+    LBrace,
+    OptionalSection(
+        RepetitiveBySectionReMaster(
+            Coma,
+            Identifier
+        )
+    ),
+    RBrace
 )
