+++ src/UnifyMain/kotlin/unify/parsers/variables/DestructuringCenterParser.kt
@@ -1,10 +1,7 @@
 package unify.parsers.variables
 
 import language.parsers.ParserStatic
-import language.sections.AlternativeSection
-import language.sections.RepetitiveBySection
-import language.sections.RepetitiveSection
-import language.sections.Section
+import language.sections.*
 import lib.matcher.TestableStatic
 import unify.tokens.characters.Coma
 import unify.tokens.characters.Dot
@@ -25,9 +22,9 @@ class DestructuringCenterParser : ParserStatic() {
                 considerNewLines = true
             )
 
-            val repetitiveBy = RepetitiveBySection(
-                OptionalMutabilityVariableDeclarationParser(),
-                Coma
+            val repetitiveBy = RepetitiveBySectionReMaster(
+                Coma,
+                OptionalMutabilityVariableDeclarationParser()
             )
 
             return arrayOf(
